move contract to lib
