+++ contracts/nzec/src/lib.rs
@@ -1,7 +1,516 @@
-#[cfg(feature = "contract")]
-mod contract;
+use near_contract_standards::{
+    fungible_token::{
+        events::{FtBurn, FtMint},
+        metadata::{FungibleTokenMetadata, FungibleTokenMetadataProvider, FT_METADATA_SPEC},
+        receiver::ext_ft_receiver,
+        resolver::ext_ft_resolver,
+        FungibleToken, FungibleTokenCore, FungibleTokenResolver,
+    },
+    storage_management::{StorageBalance, StorageBalanceBounds, StorageManagement},
+};
+use near_plugins::{events::AsEvent, only, ownable::OwnershipTransferred, Ownable};
+use near_sdk::{
+    assert_one_yocto,
+    borsh::{BorshDeserialize, BorshSerialize},
+    env,
+    json_types::{Base64VecU8, U128},
+    log, near, require,
+    store::Lazy,
+    AccountId, BorshStorageKey, Gas, NearToken, PanicOnDefault, Promise, PromiseOrValue, PublicKey,
+};
 
 pub const WITHDRAW_MEMO_PREFIX: &str = "WITHDRAW_TO:";
+
+const GAS_FOR_RESOLVE_TRANSFER: Gas = Gas::from_tgas(5);
+const GAS_FOR_FT_TRANSFER_CALL: Gas = Gas::from_tgas(30);
+
+const OUTER_UPGRADE_GAS: Gas = Gas::from_tgas(15);
+const NO_DEPOSIT: NearToken = NearToken::from_yoctonear(0);
+
+#[derive(BorshSerialize, BorshStorageKey)]
+#[borsh(crate = "::near_sdk::borsh")]
+enum Prefix {
+    FungibleToken,
+    Metadata,
+}
+
+#[derive(BorshDeserialize)]
+#[borsh(crate = "::near_sdk::borsh")]
+pub struct ContractV0 {
+    token: FungibleToken,
+    metadata: Lazy<FungibleTokenMetadata>,
+}
+
+#[near(serializers = [json])]
+pub struct PostAction {
+    pub receiver_id: AccountId,
+    pub amount: U128,
+    pub memo: Option<String>,
+    pub msg: String,
+    pub gas: Option<Gas>,
+}
+
+#[near(
+    contract_state,
+    contract_metadata(
+        standard(standard = "nep141", version = "1.0.0"),
+        standard(standard = "nep145", version = "1.0.0"),
+        standard(standard = "nep148", version = "1.0.0"),
+    )
+)]
+#[derive(Ownable, PanicOnDefault)]
+pub struct Contract {
+    bridge_id: AccountId,
+    token: FungibleToken,
+    metadata: Lazy<FungibleTokenMetadata>,
+}
+
+#[near]
+impl Contract {
+    #[init]
+    pub fn new(
+        owner_id: Option<AccountId>,
+        bridge_id: AccountId,
+        name: String,
+        symbol: String,
+        icon: Option<String>,
+        decimals: u8,
+    ) -> Self {
+        require!(!env::state_exists(), "Already initialized");
+        let mut contract = Self {
+            bridge_id,
+            token: FungibleToken::new(Prefix::FungibleToken),
+            metadata: Lazy::new(
+                Prefix::Metadata,
+                FungibleTokenMetadata {
+                    spec: FT_METADATA_SPEC.to_string(),
+                    name,
+                    symbol,
+                    icon,
+                    reference: None,
+                    reference_hash: None,
+                    decimals,
+                },
+            ),
+        };
+
+        contract
+            .token
+            .internal_register_account(&contract.bridge_id);
+
+        let owner = owner_id.unwrap_or_else(env::predecessor_account_id);
+        // Ownable::owner_set requires it to be a promise
+        require!(!env::storage_write(
+            contract.owner_storage_key(),
+            owner.as_bytes()
+        ));
+        OwnershipTransferred {
+            previous_owner: None,
+            new_owner: Some(owner),
+        }
+        .emit();
+        contract
+    }
+}
+
+#[near]
+impl Contract {
+    #[only(owner)]
+    #[payable]
+    pub fn set_metadata(
+        &mut self,
+        name: Option<String>,
+        symbol: Option<String>,
+        reference: Option<String>,
+        reference_hash: Option<Base64VecU8>,
+        decimals: Option<u8>,
+        icon: Option<String>,
+    ) {
+        assert_one_yocto();
+
+        let mut metadata = self.ft_metadata();
+        if let Some(name) = name {
+            metadata.name = name;
+        }
+        if let Some(symbol) = symbol {
+            metadata.symbol = symbol;
+        }
+        if let Some(reference) = reference {
+            metadata.reference = Some(reference);
+        }
+        if let Some(reference_hash) = reference_hash {
+            metadata.reference_hash = Some(reference_hash);
+        }
+        if let Some(decimals) = decimals {
+            // Decimals can't be changed if it's already set.
+            require!(metadata.decimals == 0, "decimals already set");
+            metadata.decimals = decimals;
+        }
+        if let Some(icon) = icon {
+            metadata.icon = Some(icon);
+        }
+
+        metadata.assert_valid();
+        self.metadata.set(metadata);
+    }
+
+    #[payable]
+    pub fn safe_mint(
+        &mut self,
+        account_id: AccountId,
+        amount: U128,
+        msg: Option<String>,
+    ) -> PromiseOrValue<U128> {
+        self.assert_bridge();
+        require!(
+            account_id != self.bridge_id,
+            "safe_mint: account_id must not be the bridge"
+        );
+        self.token.internal_deposit(&self.bridge_id, amount.into());
+
+        if self.token.accounts.get(&account_id).is_none() {
+            return PromiseOrValue::Value(U128(0));
+        }
+
+        if let Some(msg) = msg {
+            self.ft_transfer_call(account_id, amount, None, msg)
+        } else {
+            self.ft_transfer(account_id, amount, None);
+            PromiseOrValue::Value(amount)
+        }
+    }
+
+    pub fn mint(
+        &mut self,
+        mint_account_id: AccountId,
+        mint_amount: U128,
+        protocol_fee: U128,
+        relayer_account_id: AccountId,
+        relayer_fee: U128,
+        post_actions: Option<Vec<PostAction>>,
+    ) {
+        self.assert_bridge();
+        self.mint_inner(&mint_account_id, mint_amount);
+        if protocol_fee.0 > 0 {
+            self.mint_inner(&self.bridge_id.clone(), protocol_fee);
+        }
+        if relayer_fee.0 > 0 {
+            self.mint_inner(&relayer_account_id, relayer_fee);
+        }
+        if let Some(post_actions) = post_actions {
+            Self::ext(env::current_account_id())
+                .handle_post_actions(mint_account_id, post_actions)
+                .detach();
+        }
+    }
+
+    pub fn migration_mint(&mut self, entries: Vec<(AccountId, U128)>) {
+        self.assert_bridge();
+        for (account_id, amount) in entries {
+            self.mint_inner(&account_id, amount);
+        }
+    }
+
+    pub fn burn(
+        &mut self,
+        burn_account_id: AccountId,
+        burn_amount: U128,
+        relayer_account_id: AccountId,
+        relayer_fee: U128,
+    ) {
+        self.assert_bridge();
+        self.token
+            .internal_withdraw(&self.bridge_id, burn_amount.into());
+        if relayer_fee.0 > 0 {
+            if self.token.accounts.get(&relayer_account_id).is_none() {
+                self.token.internal_register_account(&relayer_account_id);
+            }
+            self.token.internal_transfer(
+                &self.bridge_id,
+                &relayer_account_id,
+                relayer_fee.into(),
+                None,
+            );
+        }
+        near_contract_standards::fungible_token::events::FtBurn {
+            owner_id: &burn_account_id,
+            amount: burn_amount,
+            memo: None,
+        }
+        .emit();
+    }
+
+    #[only(self, owner)]
+    #[payable]
+    pub fn ft_deposit(&mut self, owner_id: AccountId, amount: U128, memo: Option<String>) {
+        self.token.storage_deposit(Some(owner_id.clone()), None);
+        self.token.internal_deposit(&owner_id, amount.into());
+        FtMint {
+            owner_id: &owner_id,
+            amount,
+            memo: memo.as_deref(),
+        }
+        .emit();
+    }
+}
+
+#[near]
+impl FungibleTokenCore for Contract {
+    #[payable]
+    fn ft_transfer(&mut self, receiver_id: AccountId, amount: U128, memo: Option<String>) {
+        // A special case we created to handle withdrawals:
+        // If the receiver id is the token contract id, we burn these tokens by calling ft_withdraw,
+        // which will reduce the balance and emit an FtBurn event.
+        if receiver_id == env::current_account_id()
+            && memo
+                .as_deref()
+                .is_some_and(|memo| memo.starts_with(WITHDRAW_MEMO_PREFIX))
+        {
+            self.ft_withdraw(&env::predecessor_account_id(), amount, memo.as_deref());
+        } else {
+            self.token.ft_transfer(receiver_id, amount, memo);
+        }
+    }
+
+    #[payable]
+    fn ft_transfer_call(
+        &mut self,
+        receiver_id: AccountId,
+        amount: U128,
+        memo: Option<String>,
+        msg: String,
+    ) -> PromiseOrValue<U128> {
+        self.token.ft_transfer_call(receiver_id, amount, memo, msg)
+    }
+
+    fn ft_total_supply(&self) -> U128 {
+        self.token.ft_total_supply()
+    }
+
+    fn ft_balance_of(&self, account_id: AccountId) -> U128 {
+        self.token.ft_balance_of(account_id)
+    }
+}
+
+#[near]
+impl FungibleTokenResolver for Contract {
+    #[private]
+    fn ft_resolve_transfer(
+        &mut self,
+        sender_id: AccountId,
+        receiver_id: AccountId,
+        amount: U128,
+    ) -> U128 {
+        let (used_amount, burned_amount) =
+            self.token
+                .internal_ft_resolve_transfer(&sender_id, receiver_id, amount);
+        if burned_amount > 0 {
+            log!("Account @{} burned {}", sender_id, burned_amount);
+        }
+        used_amount.into()
+    }
+}
+
+#[near]
+impl StorageManagement for Contract {
+    #[payable]
+    fn storage_deposit(
+        &mut self,
+        account_id: Option<AccountId>,
+        registration_only: Option<bool>,
+    ) -> StorageBalance {
+        self.token.storage_deposit(account_id, registration_only)
+    }
+
+    #[payable]
+    fn storage_withdraw(&mut self, amount: Option<NearToken>) -> StorageBalance {
+        self.token.storage_withdraw(amount)
+    }
+
+    #[payable]
+    fn storage_unregister(&mut self, force: Option<bool>) -> bool {
+        #[allow(unused_variables)]
+        if let Some((account_id, balance)) = self.token.internal_storage_unregister(force) {
+            log!("Closed @{} with {}", account_id, balance);
+            true
+        } else {
+            false
+        }
+    }
+
+    fn storage_balance_bounds(&self) -> StorageBalanceBounds {
+        self.token.storage_balance_bounds()
+    }
+
+    fn storage_balance_of(&self, account_id: AccountId) -> Option<StorageBalance> {
+        self.token.storage_balance_of(account_id)
+    }
+}
+
+#[near]
+impl FungibleTokenMetadataProvider for Contract {
+    fn ft_metadata(&self) -> FungibleTokenMetadata {
+        self.metadata.clone()
+    }
+}
+
+impl Contract {
+    fn ft_withdraw(&mut self, account_id: &AccountId, amount: U128, memo: Option<&str>) {
+        assert_one_yocto();
+        require!(amount.0 > 0, "zero amount");
+        self.token.internal_withdraw(account_id, amount.into());
+        FtBurn {
+            owner_id: account_id,
+            amount,
+            memo,
+        }
+        .emit();
+    }
+}
+
+#[near]
+impl Contract {
+    #[only(self, owner)]
+    #[payable]
+    pub fn add_full_access_key(&mut self, public_key: PublicKey) -> Promise {
+        assert_one_yocto();
+        Promise::new(env::current_account_id()).add_full_access_key(public_key)
+    }
+
+    #[only(self, owner)]
+    #[payable]
+    pub fn delete_key(&mut self, public_key: PublicKey) -> Promise {
+        assert_one_yocto();
+        Promise::new(env::current_account_id()).delete_key(public_key)
+    }
+}
+
+impl Contract {
+    fn assert_bridge(&self) {
+        require!(self.bridge_id == env::predecessor_account_id(), "Not Allow");
+    }
+
+    fn mint_inner(&mut self, account_id: &AccountId, amount: U128) {
+        if self.token.accounts.get(account_id).is_none() {
+            self.token.internal_register_account(account_id);
+        }
+        self.token.internal_deposit(account_id, amount.into());
+        near_contract_standards::fungible_token::events::FtMint {
+            owner_id: account_id,
+            amount,
+            memo: None,
+        }
+        .emit();
+    }
+}
+
+#[near]
+impl Contract {
+    #[private]
+    pub fn handle_post_actions(&mut self, sender_id: AccountId, post_actions: Vec<PostAction>) {
+        for post_action in post_actions {
+            let PostAction {
+                receiver_id,
+                amount,
+                memo,
+                msg,
+                gas,
+            } = post_action;
+            if let Some(gas) = gas {
+                Self::ext(env::current_account_id())
+                    .with_static_gas(gas)
+                    .handle_post_action(sender_id.clone(), receiver_id, amount, memo, msg)
+                    .detach();
+            } else {
+                Self::ext(env::current_account_id())
+                    .handle_post_action(sender_id.clone(), receiver_id, amount, memo, msg)
+                    .detach();
+            }
+        }
+    }
+
+    #[private]
+    pub fn handle_post_action(
+        &mut self,
+        sender_id: AccountId,
+        receiver_id: AccountId,
+        amount: U128,
+        memo: Option<String>,
+        msg: String,
+    ) {
+        require!(
+            env::prepaid_gas() > GAS_FOR_FT_TRANSFER_CALL,
+            "More gas is required"
+        );
+        require!(
+            receiver_id != self.bridge_id,
+            "handle_post_action: receiver_id must not be the bridge"
+        );
+        let amount = amount.into();
+        self.token
+            .internal_transfer(&sender_id, &receiver_id, amount, memo);
+        let receiver_gas = env::prepaid_gas()
+            .checked_sub(GAS_FOR_FT_TRANSFER_CALL)
+            .unwrap_or_else(|| env::panic_str("Prepaid gas overflow"));
+        // Initiating receiver's call and the callback
+        ext_ft_receiver::ext(receiver_id.clone())
+            .with_static_gas(receiver_gas)
+            .ft_on_transfer(sender_id.clone(), amount.into(), msg)
+            .then(
+                ext_ft_resolver::ext(env::current_account_id())
+                    .with_static_gas(GAS_FOR_RESOLVE_TRANSFER)
+                    .ft_resolve_transfer(sender_id, receiver_id, amount.into()),
+            )
+            .detach();
+    }
+}
+
+#[near]
+impl Contract {
+    pub fn version(&self) -> String {
+        env!("CARGO_PKG_VERSION").to_owned()
+    }
+
+    pub fn bridge_id(&self) -> AccountId {
+        self.bridge_id.clone()
+    }
+
+    #[private]
+    #[init(ignore_state)]
+    pub fn migrate(bridge_id: AccountId) -> Self {
+        let old: ContractV0 =
+            env::state_read().unwrap_or_else(|| env::panic_str("ERR_FAILED_TO_READ_STATE"));
+        Self {
+            bridge_id,
+            token: old.token,
+            metadata: old.metadata,
+        }
+    }
+
+    #[only(owner)]
+    pub fn upgrade_and_migrate(&self) {
+        // Receive the code directly from the input to avoid the
+        // GAS overhead of deserializing parameters
+        let code = env::input().unwrap_or_else(|| env::panic_str("ERR_NO_INPUT"));
+
+        // Deploy the contract code.
+        let promise_id = env::promise_batch_create(&env::current_account_id());
+        env::promise_batch_action_deploy_contract(promise_id, &code);
+
+        // Call promise to migrate the state.
+        // Batched together to fail upgrade if migration fails.
+        env::promise_batch_action_function_call(
+            promise_id,
+            "migrate",
+            b"",
+            NO_DEPOSIT,
+            env::prepaid_gas()
+                .saturating_sub(env::used_gas())
+                .saturating_sub(OUTER_UPGRADE_GAS),
+        );
+        env::promise_return(promise_id);
+    }
+}
 
 pub fn withdraw_to(address: impl AsRef<str>) -> String {
     format!("{WITHDRAW_MEMO_PREFIX}{}", address.as_ref())
